requirements.txt
discord-ext-slash>=0.8.1